[FIX] - Se agrega HomeEntity con los modelos de negocio
